Pet CRUD with Exceptions

+++ Petshop.Core/ApplicationServices/Impl/PetService.cs
@@ -23,20 +23,32 @@ namespace Petshop.Core.ApplicationServices.Impl
 
         public Pet Read(int id)
         {
-            return _petRepository.Read(id);
+            if (id < 1) throw new ArgumentException($"Id must be a positive integer.");
+            Pet pet = _petRepository.Read(id);
+            if (pet is null) throw new KeyNotFoundException($"No record with id {id} was found.");
+            return pet;
         }
 
         public Pet Create(Pet pet)
         {
+            ValidatePetData(pet);
             return _petRepository.Create(pet);
         }
         public Pet Update(Pet pet)
         {
+            if (pet.Id < 1) throw new ArgumentException($"Id must be a positive integer.");
+            ValidatePetData(pet);
+            Pet petUpdated = _petRepository.Update(pet);
+            if (petUpdated is null) throw new KeyNotFoundException($"No record with id {pet.Id} was found.");
+
             return _petRepository.Update(pet);
         }
 
         public Pet Delete(int id)
         {
+            if (id < 1) throw new ArgumentException($"Id must be a positive integer.");
+            Pet pet = _petRepository.Delete(id);
+            if (pet is null) throw new KeyNotFoundException($"No record with id {id} was found.");
             return _petRepository.Delete(id);
         }
 
@@ -54,5 +66,16 @@ namespace Petshop.Core.ApplicationServices.Impl
         {
             return _petRepository.TopXCheapest(pets, amount);
         }
-    }
+
+        private void ValidatePetData(Pet pet)
+        {
+            if (string.IsNullOrEmpty(pet.Name) || pet.Name.Length< 2) throw new ArgumentException("Name must be minimum 2 characters long.");
+            if (!Enum.IsDefined(typeof(PetType), pet.Type)) throw new ArgumentException("The pet type is invalid.");
+            if (!DateTime.TryParse(pet.BirthDate.ToString(), out _)) throw new ArgumentException("Birth date invalid format."); // service check - api controller validates date first
+            if (DateTime.Compare(pet.BirthDate, DateTime.Now) > 0) throw new ArgumentException("Birth date must be in the past.");
+            if (!DateTime.TryParse(pet.SoldDate.ToString(), out _)) throw new ArgumentException("Sold date invalid format."); // service check - api controller validates date first
+            if (DateTime.Compare(pet.SoldDate, DateTime.Now) > 0) throw new ArgumentException("Sold date must be in the past.");
+            if (pet.Price< 0) throw new ArgumentException("Price must be an integer greater than or equal to 0.");
+        }
+}
 }
